refactor pipeline
